+++ src/line.rs
@@ -36,7 +36,7 @@ impl Line {
 
         // no intersection found
         if determinant == 0.0 {
-            return Float2::new(f64::MAX, f64::MAX);
+            return Float2::new(f32::MAX, f32::MAX);
         } else {
             let x = (b2 * c1 - b1 * c2) / determinant;
             let y = (a1 * c2 - a2 * c1) / determinant;
